hub/bridge/requirements.txt
aiomqtt==2.5.1
asyncpg==0.31.0
pydantic==2.12.5
python-dotenv
aiohttp==3.11.18
